Adding null check on gameimag file.move.

+++ SpeedRunAppImport.Service/GameService.cs
@@ -415,9 +415,12 @@ namespace SpeedRunAppImport.Service
             {
                 var fileName = Path.GetFileName(tempGameCoverPath.Value);
                 var destFilePath = Path.Combine(BaseWebPath, GameImageWebPath, fileName);
-                File.Move(tempGameCoverPath.Value, destFilePath, true);
-                var gameCoverPath = Path.Combine("/" + GameImageWebPath, fileName);
-                gameCoverPaths.Add(tempGameCoverPath.Key, gameCoverPath);
+                if (File.Exists(tempGameCoverPath.Value))
+                {
+                    File.Move(tempGameCoverPath.Value, destFilePath, true);
+                    var gameCoverPath = Path.Combine("/" + GameImageWebPath, fileName);
+                    gameCoverPaths.Add(tempGameCoverPath.Key, gameCoverPath);
+                }
             }
 
             return gameCoverPaths;
